requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2.3
djangorestframework==3.12.4
greenlet==1.1.0
idna==2.10
jwt==1.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.17
SQLAlchemy-Utils==0.37.5
sqlparse==0.4.1
urllib3==1.26.5
gunicorn==20.1.0
